% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progressList.R
\name{progressList}
\alias{progressList}
\title{Create a new progressList}
\usage{
progressList(inputId, label, status)
}
\arguments{
\item{inputId}{The input id of the widget}

\item{label}{A list of labels - one for each step in the process}

\item{status}{A list of status flags - one for each step in the process}
}
\description{
Create a new progressList
}

