% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progressList.R
\name{progUpdate}
\alias{progUpdate}
\title{Update a progress flag to a new state}
\usage{
progUpdate(session, value, newstatus)
}
\arguments{
\item{session}{The session object passed to function given to \code{shinyServer}}

\item{value}{Text string of the status to update}

\item{newstatus}{Updated status}
}
\description{
Update a progress flag to a new state
}

