% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progressList.R
\name{progUpdateList}
\alias{progUpdateList}
\title{Update multiple flags to new states}
\usage{
progUpdateList(session, label, status)
}
\arguments{
\item{session}{The session object passed to function given to \code{shinyServer}}

\item{label}{List of labels to update}

\item{status}{List of new states}
}
\description{
Update multiple flags to new states
}

